applications/utils/requirements-bazel.txt
pytest==7.2.2
pytest-asyncio==0.21.0
python-dateutil==2.8.2
paho-mqtt==1.6.1
pyftpdlib==1.5.7
ocpp==0.17.0
websockets==10.4
pyOpenSSL==23.0.0
cryptography==39.0.2
packaging==23.0
pycparser==2.21
attrs>=22.2.0
cffi==1.15.1
exceptiongroup==1.1.1
iniconfig==2.0.0
jsonschema==4.17.3
jsonschema-specifications==2023.11.1
pluggy==1.0.0
pyrsistent==0.19.3
PyYAML==6.0.1
MarkupSafe==2.1.3
six==1.16.0
tomli==2.0.1
Jinja2==3.1.2
referencing==0.31.0
rpds-py==0.13.1
stringcase==1.2.0
pkgutil_resolve_name>=1.3.10
importlib_resources>=1.4.0
zipp>=1.0.0
pyyaml==6.0.1
